Changing gid and uid from hardcoded values to dynamic ones.
The hardcoded values are in the user range and my collide with users
on the system. The group and user are also being setup at the system
level and should be in that range instead of the user range.

Change-Id: I66ac095cd94c998a721e26fc88d7c017866b09ea
(cherry picked from commit e2dd1ba138886d1834b0504cde691d1a79b2cfe7)